backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
black==25.12.0
cachetools==6.2.2
celery==5.3.6
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
deepseek-cli==0.2.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-nested-routers==0.95.0
drf-spectacular==0.27.0
et_xmlfile==2.0.0
flake8==7.3.0
flake8-docstrings==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
isort==7.0.0
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pgvector==0.4.2
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.12.0
pydantic_core==2.41.1
pydocstyle==6.3.0
pyfiglet==1.0.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.27
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
